openface-service/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
numpy==1.24.4
opencv-python==4.12.0.88
pandas==2.0.3
protobuf==5.29.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
six==1.17.0
tzdata==2025.2
websockets==13.1
zmq==0.0.0
